feat: 댓글 상태 변경 api 스펙 변경에 따른 대응
