+++ android/app/src/main/java/com/zzang/chongdae/data/remote/api/CommentApiService.kt
@@ -4,9 +4,11 @@ import com.zzang.chongdae.data.remote.dto.request.CommentRequest
 import com.zzang.chongdae.data.remote.dto.response.CommentRoomsResponse
 import com.zzang.chongdae.data.remote.dto.response.CommentsResponse
 import com.zzang.chongdae.data.remote.dto.response.CommentOfferingInfoResponse
+import com.zzang.chongdae.data.remote.dto.response.UpdatedStatusResponse
 import retrofit2.Response
 import retrofit2.http.Body
 import retrofit2.http.GET
+import retrofit2.http.PATCH
 import retrofit2.http.POST
 import retrofit2.http.Query
 
@@ -28,4 +30,9 @@ interface CommentApiService {
     suspend fun getCommentOfferingInfo(
         @Query("offering-id") offeringId: Long,
     ): Response<CommentOfferingInfoResponse>
+
+    @PATCH("/comments/status")
+    suspend fun patchOfferingStatus(
+        @Query("offering-id") offeringId: Long,
+    ): Response<UpdatedStatusResponse>
 }
