+++ android/app/src/main/java/com/zzang/chongdae/data/remote/api/OfferingApiService.kt
@@ -12,7 +12,6 @@ import retrofit2.Response
 import retrofit2.http.Body
 import retrofit2.http.GET
 import retrofit2.http.Multipart
-import retrofit2.http.PATCH
 import retrofit2.http.POST
 import retrofit2.http.Part
 import retrofit2.http.Path
@@ -55,9 +54,4 @@ interface OfferingApiService {
     suspend fun postProductImageS3(
         @Part image: MultipartBody.Part,
     ): Response<ProductUrlResponse>
-
-    @PATCH("/offerings/{offering-id}/status")
-    suspend fun patchOfferingStatus(
-        @Path("offering-id") offeringId: Long,
-    ): Response<Unit>
 }
