+++ android/app/src/main/java/com/zzang/chongdae/data/remote/source/CommentRemoteDataSourceImpl.kt
@@ -17,7 +17,7 @@ class CommentRemoteDataSourceImpl(
             if (response.isSuccessful) {
                 response.body() ?: error("에러 발생: null")
             } else {
-                error("$${ERROR_PREFIX}{response.code()}")
+                error("${ERROR_PREFIX}${response.code()}")
             }
         }
 
@@ -36,6 +36,17 @@ class CommentRemoteDataSourceImpl(
         return runCatching {
             val response: Response<CommentOfferingInfoResponse> =
                 service.getCommentOfferingInfo(offeringId)
+            if (response.isSuccessful) {
+                response.body() ?: error(ERROR_PREFIX + ERROR_NULL_MESSAGE)
+            } else {
+                error("${ERROR_PREFIX}${response.code()}")
+            }
+        }
+    }
+    
+    override suspend fun updateOfferingStatus(offeringId: Long): Result<Unit> {
+        return runCatching {
+            val response = service.patchOfferingStatus(offeringId)
             if (response.isSuccessful) {
                 response.body() ?: error(ERROR_PREFIX + ERROR_NULL_MESSAGE)
             } else {
