+++ android/app/src/main/java/com/zzang/chongdae/data/repository/CommentDetailRepositoryImpl.kt
@@ -67,4 +67,8 @@ class CommentDetailRepositoryImpl(
             commentOfferingInfo.toDomain()
         }
     }
+    
+    override suspend fun updateOfferingStatus(offeringId: Long): Result<Unit> {
+        return commentRemoteDataSource.updateOfferingStatus(offeringId)
+    }
 }
